atienzisaa COMPLETION
Issue: camera is asymmetric therefore, renders me unable to keep the robot on the middle of the white line once its moved. Robot thinks that if line is on right side its fine since this is default. (4 pixels on right and 0 on left is default).
Fix: changing the robot.hpp file and moving the width of the camera by 2 pixels to the left to balance it out and make it symmetrical if possible.